+++ src/components/CardPassword/index.tsx
@@ -1,5 +1,5 @@
-import React, { useEffect, useState } from 'react';
 import CardPasswordView from './CardPasswordView';
+import { useCardPassword } from '../../hooks/useCardPassword';
 
 export interface CardPasswordProps {
   passwordNumbers: string;
@@ -7,33 +7,15 @@ export interface CardPasswordProps {
   setPasswordError: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-const PASSWORD_NUMBERS_LENGTH = 2;
-const ERROR_MESSAGE = '숫자 2자리를 입력해주세요.';
-
 const CardPassword = ({
   passwordNumbers,
   setPasswordNumbers,
   setPasswordError,
 }: CardPasswordProps) => {
-  const [errorMessage, setErrorMessage] = useState('');
-  const [error, setError] = useState(false);
-
-  useEffect(() => {
-    const hasError = error;
-    setPasswordError(hasError);
-  }, [error]);
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const value = e.target.value;
-    if (/^[0-9]*$/.test(value) && value.length <= PASSWORD_NUMBERS_LENGTH) {
-      setErrorMessage('');
-      setError(false);
-      setPasswordNumbers(value);
-    } else {
-      setErrorMessage(ERROR_MESSAGE);
-      setError(true);
-    }
-  };
+  const { errorMessage, error, handleInputChange } = useCardPassword({
+    setPasswordNumbers,
+    setPasswordError,
+  });
 
   return (
     <CardPasswordView
